Start adding users by username
