+++ src/flick/api/urls.py
@@ -14,7 +14,8 @@ from friend.views import (
     FriendRejectListAndCreate,
     FriendRemoveListAndCreate,
 )
-from show.views import SearchShow, ShowViewSet
+from search.views import Search
+from show.views import ShowViewSet
 from tag.views import TagList, TagDetail
 
 router = routers.DefaultRouter()
@@ -41,7 +42,7 @@ urlpatterns = [
     path("items/<int:pk>/", ItemDetail.as_view(), name="item-detail"),
     path("like/", LikeItem.as_view(), name="like"),
     path("media/image/", UploadImage.as_view(), name="upload"),
-    path("search/", SearchShow.as_view(), name="search-show"),
+    path("search/", Search.as_view(), name="search-show"),
     path("tags/", TagList.as_view(), name="tag-list"),
     path("tags/<int:pk>/", TagDetail.as_view(), name="tag-detail"),
 ]
